# docs/openapi.yaml
openapi: 3.1.0
info:
  title: ClipFlow API
  version: 0.1.0
servers:
  - url: http://localhost:8080
paths:
  /v1/jobs:
    post:
      summary: Validate and durably enqueue one MP4
      parameters:
        - in: header
          name: Idempotency-Key
          required: true
          schema: {type: string, minLength: 1, maxLength: 200}
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              additionalProperties: false
              properties:
                file: {type: string, format: binary, description: MP4 up to 20 MiB and 30 seconds}
      responses:
        '202': {description: New or original durable job, content: {application/json: {schema: {$ref: '#/components/schemas/Job'}}}}
        '400': {$ref: '#/components/responses/Error'}
        '409': {$ref: '#/components/responses/Error'}
        '413': {$ref: '#/components/responses/Error'}
        '422': {$ref: '#/components/responses/Error'}
        '503': {$ref: '#/components/responses/Error'}
    get:
      summary: Most recent 50 jobs, newest first
      responses:
        '200': {description: Job list, content: {application/json: {schema: {type: array, items: {$ref: '#/components/schemas/Job'}}}}}
        '503': {$ref: '#/components/responses/Error'}
  /v1/jobs/{id}:
    parameters:
      - {$ref: '#/components/parameters/JobID'}
    get:
      summary: Job state, metadata, output links and attempt history
      responses:
        '200': {description: Job details, content: {application/json: {schema: {$ref: '#/components/schemas/Job'}}}}
        '400': {$ref: '#/components/responses/Error'}
        '404': {$ref: '#/components/responses/Error'}
        '503': {$ref: '#/components/responses/Error'}
  /v1/jobs/{id}/outputs/{kind}:
    parameters:
      - {$ref: '#/components/parameters/JobID'}
      - in: path
        name: kind
        required: true
        schema: {type: string, enum: [thumbnail, preview, metadata]}
    get:
      summary: Proxy a published object; unpublished attempts are inaccessible
      responses:
        '200':
          description: Published object
          content:
            image/jpeg: {schema: {type: string, format: binary}}
            video/mp4: {schema: {type: string, format: binary}}
            application/json: {schema: {$ref: '#/components/schemas/Manifest'}}
        '400': {$ref: '#/components/responses/Error'}
        '404': {$ref: '#/components/responses/Error'}
        '409': {$ref: '#/components/responses/Error'}
        '503': {$ref: '#/components/responses/Error'}
  /healthz:
    get:
      summary: Process health
      responses:
        '200': {description: Process alive}
  /readyz:
    get:
      summary: Database and bucket connectivity
      responses:
        '200': {description: Ready}
        '503': {$ref: '#/components/responses/Error'}
  /metrics:
    get:
      summary: Durable Prometheus queue/outcome counters and timing summaries
      responses:
        '200': {description: Metrics, content: {text/plain: {schema: {type: string}}}}
        '503': {$ref: '#/components/responses/Error'}
components:
  parameters:
    JobID:
      in: path
      name: id
      required: true
      schema: {type: string, format: uuid}
  responses:
    Error:
      description: Request or service error
      content:
        application/json:
          schema:
            type: object
            required: [error]
            properties:
              error:
                type: object
                required: [code, message]
                properties:
                  code: {type: string}
                  message: {type: string}
  schemas:
    Media:
      type: object
      required: [duration, width, height, size, audio]
      properties:
        duration: {type: number}
        width: {type: integer}
        height: {type: integer}
        size: {type: integer}
        audio: {type: boolean}
    Manifest:
      type: object
      required: [objects, input, preview, sizes]
      properties:
        objects: {type: object, additionalProperties: {type: string}}
        input: {$ref: '#/components/schemas/Media'}
        preview: {$ref: '#/components/schemas/Media'}
        sizes: {type: object, additionalProperties: {type: integer}}
    Attempt:
      type: object
      properties:
        number: {type: integer}
        token: {type: string, format: uuid}
        worker: {type: string}
        started_at: {type: string, format: date-time}
        ended_at: {type: [string, 'null'], format: date-time}
        outcome: {type: string}
        error_category: {type: [string, 'null']}
        error_message: {type: [string, 'null']}
        output_prefix: {type: string}
    Job:
      type: object
      required: [id, filename, state, attempt_count, created_at, updated_at, attempts]
      properties:
        id: {type: string, format: uuid}
        filename: {type: string}
        state: {type: string, enum: [queued, running, succeeded, failed]}
        attempt_count: {type: integer}
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
        completed_at: {type: [string, 'null'], format: date-time}
        error_category: {type: [string, 'null']}
        error_message: {type: [string, 'null']}
        metadata: {$ref: '#/components/schemas/Manifest'}
        outputs: {type: object, additionalProperties: {type: string}}
        attempts: {type: array, items: {$ref: '#/components/schemas/Attempt'}}